fixed kafka issues and started working on creating budgets feature

+++ savvy-app-backend/transaction-service/src/main/kotlin/com/savvy/transactionservice/dto/CategoryDTO.kt
@@ -1,3 +1,3 @@
 package com.savvy.transactionservice.dto
 
-data class CategoryDTO(val id: Int, val name: String)
+data class CategoryDTO(val id: Long, val name: String)
